HSEARCH-3115 Move Lucene analysis DSL context delegation code to an abstract parent class
We're about to add more methods that we will have to delegate to, so
let's avoid copy/pasting this delegation code everywhere.
